SVSM: Ensure VMPL0 is running with proper SEV features
Make sure the SVSM (VMPL0) is executed with required and supported SEV
features. Right now, this entails making sure it runs with restricted
injection and SNPActive, and that it doesn't run with features that the
SVSM does not support. Restricted injection is enforced to further reduce
the security exposure within the SVSM. There are also SEV features for
which the SVSM doesn't need to worry about.

These checks should change when other features are supported.

Signed-off-by: Carlos Bilbao <carlos.bilbao@amd.com>