SVSM: Ensure VMPL1 is running with proper SEV features
Make sure the guest (VMPL1) is executed with required and supported SEV
features. Right now, this entails making sure it runs with SNPActive, and
that it doesn't run with features that the SVSM does not support (for
example, VMSA Register Protection). There are also SEV features for which
the SVSM doesn't need to worry about.

These checks should change when other features are supported.

Signed-off-by: Carlos Bilbao <carlos.bilbao@amd.com>