scripts: Improve launch checks for memory/disk extensions
Some users have seen complaints from the launch script regarding the checks
performed to figure out if the memory value was specified in MBs or GBs.
Change these checks for others with better compatibility. Do the same for
the checks on file format, adding a new variable to simplify code.

Signed-off-by: Carlos Bilbao <carlos.bilbao@amd.com>