SVSM: Do not allow Secure TSC feature in the guest
The SVSM must be able to set the VMSA fields related to Secure TSC when
the guest has enabled the SecureTSC feature in SEV_FEATURES. This is
required in order to support live migration, as the VMSA Secure TSC fields
can be different between platforms and so the SVSM must be responsible for
them.

Until the SVSM has such support, do not allow the SecureTSC feature for a
guest.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>