SVSM: CPU: Maintain per-CPU TSS
Add support for a per-CPU TSS, which is needed to support IST stacks for
exception handling.

For the BSP, the TSS is updated in the GDT and then loaded to set the task
register (TR). For APs, the TR can be set directly in the VMSA when
creating the AP.

Signed-off-by: Carlos Bilbao <carlos.bilbao@amd.com>