pgtable: Remove the zeroing of memory during a page state change
Currently when changing page states, the memory is cleared. This is done
because the data itself is not altered, only the view of it. As such, the
data would now look like ciphertext. However, it should be up to the
caller to decide whether to clear the memory.

Remove the clearing of memory during a page state change and update any
call sites to clear the memory (if needed).

Fixes: 0339ac2f12d6 ("SVSM: Initial commit of Linux SVSM PoC")
Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>