README: better way to handle ovmf python not found
In systems (e.g. Ubuntu 22.04) where python is not aliased
as python2 or python3, scripts using python will
not find the executable.

One way to handle it is to install python-is-python3. However,
this may break the system and need great care from the user.

The original workaround is similar to the way above, but the
copied binary is not managed by the package manager.

Another way to fix it is to explicitly use python3 or python2
in the script, which has less impact on the system.

This commit introduces a way to fix the building script.

Signed-off-by: Hongren Zheng <i@zenithal.me>
Reviewed-by: Carlos Bilbao <carlos.bilbao@amd.com>