SVSM: Provide specific types and access functions for external symbols
The external symbols are currently just defined as u64 entries. Provide
functions to access them so that type checking can be properly performed
and the symbol type is better documented.

This fixes a number of issues where the symbol is actually a virtual
address, as opposed to a physical address, that would cause an issue if
an offset page table is used.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>
Reviewed-by: Carlos Bilbao <carlos.bilbao@amd.com>