Add rust wrappers required by external dependencies
The libcrt implements the libc interface required to build SVSM external
dependencies, but part of it is backed by SVSM core functions
implemented in Rust.

This patch adds wrappers for malloc(), calloc(), realloc(), free(),
abort() and serial_out(). The latter is not part of the libc standard, but
the libcrt uses it as a backend for printf.

The "-Wl,-u,malloc" is used in the Makefile to prevent the linker from
removing the wrapper.rs symbols.

Co-developed-by: Vikram Narayanan <vikram186@gmail.com>
Signed-off-by: Vikram Narayanan <vikram186@gmail.com>
Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>