Makefile: Add target to configure openssl
Add a Makefile target to configure openssl for the attestation
report feature.

Why are we using openssl-1.1.1q?
        * It has smaller memory footprint than the openssl 3.
        * Later we will add vTPM functionalities to the SVSM and the
Microsoft TPM 2.0 reference implementation (ms-tpm-20-ref) recommends
the openssl 1.1.1d or higher.

How do we configure openssl?
        * Similar to how openssl is configured in OVMF.
        * Disable the openssl components that are not required to
protect the communication channel with the PSP.
        * Don't select any rand seed source (--with-rand-seed=none).
Currently, we don't use DRBG, so that's fine. If we need DRBG in the
future, we can add wrappers for the openssl functions that populate the
entropy pool and call them; or we can configure openssl with
"--with-rand-seed=rdcpu", which tells openssl to use RDRAND and RDSEED
to populate the entropy pool automatically.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>