mem: Add MapGuard to automatically unmap areas
MapGuard maps a physical address range into virtual memory and returns a
guard object.  When it is dropped (goes out of scope), MapGuard will
unmap the area (unless it was constructed with
`new_private_persistent()`).

This allows returning from middle of scopes (using `return` or `?`)
without keeping manual track of which areas should be unmapped.

Signed-off-by: Dov Murik <dovmurik@linux.ibm.com>