Makefile: Add target to build libcrt
The libcrt.a library is included in the SVSM binary.

The libcrt Makefile is always called to decide when the library needs to
be rebuilt.

External libraries don't need to be cleaned very often. The clean_all
target added in this patch can be used to clean all objects, including
the external libraries'.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>