wrapper: Use libc functions when testing
When compiling and executing under normal conditions, custom wrappers are
provided for a number of libc memory operations, such as free() and alloc().
However, when running `make test`, naming conflicts with standard libc
cause the test to fail. Make wrapper functions configuration-dependent, and
use libc when testing. This ensures that the correct functions are used
under normal conditions and during testing. Fixes issue #45.

Fixes: da5be6086d1b ("Add rust wrappers required by external dependencies")
Signed-off-by: Carlos Bilbao <carlos.bilbao@amd.com>