alloc: Remove alloc_error_handler
The experimental alloc_error_handler feature was removed in Rust 1.71.0.
Starting from 1.71.0, Rust will panic when there's an allocation error
(OOM), which is the expected behaviour in SVSM (it will call the
SVSM-defined panic handler).

Remove alloc_error_handler mentions, and declare 1.71 as the minimal
Rust version for building linux-svsm.

Signed-off-by: Dov Murik <dovmurik@linux.ibm.com>