build: Use bindgen to build FFI bindings
Add bindgen to automatically produce Rust FFI (Foreign Functions
Interface) bindings out of C headers to be used in other crates. Cargo
picks up on the existence of build.rs, then compile and execute it
before the rest of the create is built. The resulting bindings are
written to src/bindings.rs.

Rename the target x86_64-unknown-none as bindgen uses it to properly
generate bindings.

Co-developed-by: Vikram Narayanan <vikram186@gmail.com>
Signed-off-by: Vikram Narayanan <vikram186@gmail.com>
Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>