psp/guest_request_cmd: Add SNP Guest Request data structures
The PSP expects each SNP Guest Request command received to be wrapped by
the SnpGuestRequestMsg structure. The command itself goes in the payload
of the message and has its own structure definition.

When the PSP receives a SnpGuestRequestMsg, it processes the command
found in the payload and sends the result back also wrapped by the
SnpGuestRequestMsg. The command result has its own structure definition
as well.

For example, for an attestation report request we wrap the
SnpReportRequest structure with the SnpGuestRequestMsg. The PSP will
unwrap the message, process the command and send back the attestatio
report result in the SnpReportResponse structure, wrapped by the
SnpGuestRequestMsg.

Other commands types can be wrapped using the SnpGuestRequestMsg.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>