psp/guest_request_cmd: Add init_ossl_ctx()
Calls to openssl usually requires a context. For example, before calling
openssl to encrypt data, first we need to create a new context and set
it up for the encryption algorithm we want to use.

Once we are done with the encryption session, we have to free the
context resources to avoid memory leaking.

In order to simplify error handling, this patch free the context
resources at the beginning of the session rather than at the tail.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>