protocols/core: Clear page after PVALIDATE of the page
A malicious hypervisor can attempt to reveal data from the SVSM to lower
VMPL levels through RMP manipulation related to page validation. For
example:

  - Initially, VMPL0 has a page at GPA A which maps to SPA X
  - VMPL3 asks HV to change the state of GPA B to private
  - HV maliciously reclaims SPA X and changes the RMP entry (and NPT) to
    map it at GPA B
  - VMPL3 asks VMPL0 to validate a new page at GPA B
  - VMPL0 PVALIDATE/RMPADJUSTs GPA B, allowing VMPL3 to read the data that
    VMPL0 had previously stored at GPA A

To prevent the exposure of any data in that page, the SVSM must zero-out
the memory after the PVALIDATE but before the RMPADJUST that grants
permission to the lower VMPL levels.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>