Makefile: Allow passing parameters to the rust installer
Add the RUST_INSTALLER_ARGS variable to allow passing parameters to the
Rust installer in the 'make prereq'. For example:

$ make prereq RUST_INSTALLER_ARGS='-y'

The cmdline above tells the rust installer that we want to go with
the default options without any prompt. It could be useful in the github
CI, where tty is not available.

If RUST_INSTALLER_ARGS is not set outside the Makefile, the Rust
installer is executed without any parameter.

All assignments to RUST_INSTALLER_ARGS within the Makefile is ignored,
unless the override directive is used.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>