crypto: Add crypto API for SNP guest messages
Add a generic crypto API required to encrypt and decrypt SNP guest
messages.

This ensures that the nossl.rs stubs are compiled in when we compile the
SVSM with the test feature flag set (e.g. cargo test). This is important
because the test binary is meant to run standalone; so, if we try to
compile any code that calls the SVSM OpenSSL, that will likely cause
compilation conflicts between the libcrt headers and the system glibc.

This also adds a placeholder to when the code is compiled without
setting the test feature flag (e.g. cargo build). Later we can replace
that to point out to a real crypto implementation.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>