psp/guest_request_cmd: Add SNP Guest Request data structures
Each SNP guest request command has its own request and response
structure definition. However, they are sent and received from the PSP
wrapped in the SnpGuestRequestMsg structure.

SNP guest request commands are communicated through the hypervisor, so
they are expected to be encrypted with a key known only by the guest and
the PSP.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>